Scribble/requirements.txt
streamlit
opencv-python
easyocr
pillow
numpy
google-genai
python-docx
reportlab
